+++ src/assets/dict/en-US.dic
