+++ src/assets/tray/mac@2x.png
