+++ src/assets/tray/macHighlight.png
