+++ src/assets/tray/macHighlight@2x.png
