+++ src/assets/tray/macTemplate.png
