+++ src/assets/tray/macTemplate@2x.png
