+++ src/assets/tray/windows.ico
