+++ src/spellcheck/spellcheck-provider.js
